Moved files to checklist folder
